add metadata to each page; make pages server components
